feat(mysql): preserve narrow integer and float types in Arrow transports

+++ connectorx/src/transports/mysql_arrow.rs
@@ -39,14 +39,14 @@ impl_transport!(
     systems = MySQLTypeSystem => ArrowTypeSystem,
     route = MySQLSource<BinaryProtocol> => ArrowDestination,
     mappings = {
-        { Float[f32]                 => Float64[f64]            | conversion auto }
+        { Float[f32]                 => Float32[f32]            | conversion auto }
         { Double[f64]                => Float64[f64]            | conversion auto }
-        { Tiny[i8]                   => Int16[i16]              | conversion auto }
+        { Tiny[i8]                   => Int8[i8]                | conversion auto }
         { Short[i16]                 => Int16[i16]              | conversion auto }
         { Int24[i32]                 => Int32[i32]              | conversion none }
         { Long[i32]                  => Int32[i32]              | conversion auto }
         { LongLong[i64]              => Int64[i64]              | conversion auto }
-        { UTiny[u8]                  => UInt16[u16]             | conversion auto }
+        { UTiny[u8]                  => UInt8[u8]               | conversion auto }
         { UShort[u16]                => UInt16[u16]             | conversion auto }
         { UInt24[u32]                => UInt32[u32]             | conversion none }
         { ULong[u32]                 => UInt32[u32]             | conversion auto }
@@ -75,14 +75,14 @@ impl_transport!(
     systems = MySQLTypeSystem => ArrowTypeSystem,
     route = MySQLSource<TextProtocol> => ArrowDestination,
     mappings = {
-        { Float[f32]                 => Float64[f64]            | conversion auto }
+        { Float[f32]                 => Float32[f32]            | conversion auto }
         { Double[f64]                => Float64[f64]            | conversion auto }
-        { Tiny[i8]                   => Int16[i16]              | conversion auto }
+        { Tiny[i8]                   => Int8[i8]                | conversion auto }
         { Short[i16]                 => Int16[i16]              | conversion auto }
         { Int24[i32]                 => Int32[i32]              | conversion none }
         { Long[i32]                  => Int32[i32]              | conversion auto }
         { LongLong[i64]              => Int64[i64]              | conversion auto }
-        { UTiny[u8]                  => UInt16[u16]             | conversion auto }
+        { UTiny[u8]                  => UInt8[u8]               | conversion auto }
         { UShort[u16]                => UInt16[u16]             | conversion auto }
         { UInt24[u32]                => UInt32[u32]             | conversion none }
         { ULong[u32]                 => UInt32[u32]             | conversion auto }
